Hw5: move code generated by jooq to edu.java.domain.jooq.generated package
