+++ pom.xml
@@ -169,6 +169,11 @@
                 <groupId>org.jacoco</groupId>
                 <artifactId>jacoco-maven-plugin</artifactId>
                 <version>${jacoco-maven-plugin.version}</version>
+                <configuration>
+                    <excludes>
+                        <exclude>**/jooq/generated/**</exclude>
+                    </excludes>
+                </configuration>
             </plugin>
 
         </plugins>
