+++ scrapper/src/main/java/edu/java/domain/jooq/SubscriptionRepositoryJooqImpl.java
@@ -3,9 +3,9 @@ package edu.java.domain.jooq;
 import edu.java.domain.SubscriptionRepository;
 import edu.java.dto.dao.LinkDto;
 import edu.java.dto.dao.SubscriptionDto;
-import edu.java.scrapper.domain.jooq.tables.Link;
-import edu.java.scrapper.domain.jooq.tables.Subscription;
-import edu.java.scrapper.domain.jooq.tables.records.SubscriptionRecord;
+import edu.java.domain.jooq.generated.tables.Link;
+import edu.java.domain.jooq.generated.tables.Subscription;
+import edu.java.domain.jooq.generated.tables.records.SubscriptionRecord;
 import java.time.OffsetDateTime;
 import java.util.List;
 import java.util.Optional;
